Added init and move interface
