+++ Node/src/main/java/com/github/krextouch/rtds/node/repository/Client.java
@@ -17,7 +17,8 @@ public class Client {
     @Id
     public short id;
 
-    public short curPos;
+    public short[] curPos;
 
-    public short destPos;
+    public short[] destPos;
+
 }
