+++ Node/src/main/java/com/github/krextouch/rtds/node/service/model/MoveRequest.java
@@ -12,6 +12,9 @@ import lombok.NoArgsConstructor;
 public class MoveRequest {
 
     private short clientId;
-    private short curPos;
-    private short destPos;
+
+    private short[] curPos;
+
+    private short[] destPos;
+
 }
