+++ Node/src/main/java/com/github/krextouch/rtds/node/service/node/NodeController.java
@@ -1,4 +1,4 @@
-package com.github.krextouch.rtds.node.service;
+package com.github.krextouch.rtds.node.service.node;
 
 import com.github.krextouch.rtds.node.service.model.MoveRequest;
 import org.springframework.web.bind.annotation.*;
